Big WIP to migrate the cassette sending code to a timer.
This is inspired by the TZX project for the ZX spectrum.

The TL;DR is that now the cassette sending is done via a timer1
ISR and the data is read from SD card.  Now, since the SD reading
doesn't happen as an ISR (which would be super convenient!),
the main runtime loop will have to be UI and SD IO, and the
timer1 ISR will drive cassette audio.

This is a big WIP as I've not yet tested it on the apple-ii;
i'm just putting the state machine together and testing
reading a single binary file from the SD card.

I'll delete the old code once I know this is actually able to
reliably send data!